Bringing DPA-A shutdown into alignment with DPA-B
Changed the DPA-A page wording to match changes made in the DPA-B page.  In addition, added in a new bullet item to the CAP creation text suggesting a wspow2a if there will be a long downtime.   (WSPOW2A link is currently broken, but am preserving its setup in expectation it will be populated in near-term.) .